+++ src/main/java/com/dt/user/model/SalesAmazonAd/SalesAmazonAdCpr.java
@@ -7,8 +7,8 @@ public class SalesAmazonAdCpr {
 
     private Long adCprId;
     private Long date;
-    private Long shopId;
-    private Long siteId;
+    private Integer shopId;
+    private Integer siteId;
     private Long skuId;
     private String advertisedSku;
     private String advertisedAsin;
@@ -42,7 +42,7 @@ public class SalesAmazonAdCpr {
 
     }
 
-    public SalesAmazonAdCpr(Long shopId, Long siteId, Long createDate, Long createIdUser, Long recordingId) {
+    public SalesAmazonAdCpr(Integer shopId, Integer siteId, Long createDate, Long createIdUser, Long recordingId) {
         this.shopId = shopId;
         this.siteId = siteId;
         this.createDate = createDate;
@@ -75,19 +75,19 @@ public class SalesAmazonAdCpr {
         this.date = date;
     }
 
-    public Long getShopId() {
+    public Integer getShopId() {
         return shopId;
     }
 
-    public void setShopId(Long shopId) {
+    public void setShopId(Integer shopId) {
         this.shopId = shopId;
     }
 
-    public Long getSiteId() {
+    public Integer getSiteId() {
         return siteId;
     }
 
-    public void setSiteId(Long siteId) {
+    public void setSiteId(Integer siteId) {
         this.siteId = siteId;
     }
