+++ src/main/java/com/dt/user/model/SalesAmazonAd/SalesAmazonAdHl.java
@@ -7,8 +7,8 @@ public class SalesAmazonAdHl {
 
   private Long id;
   private Long date;
-  private Long shopId;
-  private Long siteId;
+  private Integer shopId;
+  private Integer siteId;
   private String campaignName;
   private Double impressions;
   private Double clicks;
@@ -35,7 +35,7 @@ public class SalesAmazonAdHl {
 
   }
 
-  public SalesAmazonAdHl(Long shopId, Long siteId, Long createDate, Long createIdUser, Long recordingId) {
+  public SalesAmazonAdHl(Integer shopId, Integer siteId, Long createDate, Long createIdUser, Long recordingId) {
     this.shopId = shopId;
     this.siteId = siteId;
     this.createDate = createDate;
@@ -59,19 +59,19 @@ public class SalesAmazonAdHl {
     this.date = date;
   }
 
-  public Long getShopId() {
+  public Integer getShopId() {
     return shopId;
   }
 
-  public void setShopId(Long shopId) {
+  public void setShopId(Integer shopId) {
     this.shopId = shopId;
   }
 
-  public Long getSiteId() {
+  public Integer getSiteId() {
     return siteId;
   }
 
-  public void setSiteId(Long siteId) {
+  public void setSiteId(Integer siteId) {
     this.siteId = siteId;
   }
