+++ src/main/java/com/dt/user/model/SalesAmazonAd/SalesAmazonAdOar.java
@@ -5,8 +5,8 @@ public class SalesAmazonAdOar {
 
   private Long id;
   private Long date;
-  private Long shopId;
-  private Long siteId;
+  private Integer shopId;
+  private Integer siteId;
   private String campaignName;
   private String adGroupName;
   private String advertisedSku;
@@ -31,7 +31,7 @@ public class SalesAmazonAdOar {
   public SalesAmazonAdOar() {
 
   }
-  public SalesAmazonAdOar(Long shopId, Long siteId, Long createDate, Long createIdUser, Long recordingId) {
+  public SalesAmazonAdOar(Integer shopId, Integer siteId, Long createDate, Long createIdUser, Long recordingId) {
     this.shopId = shopId;
     this.siteId = siteId;
     this.createDate = createDate;
@@ -55,19 +55,19 @@ public class SalesAmazonAdOar {
     this.date = date;
   }
 
-  public Long getShopId() {
+  public Integer getShopId() {
     return shopId;
   }
 
-  public void setShopId(Long shopId) {
+  public void setShopId(Integer shopId) {
     this.shopId = shopId;
   }
 
-  public Long getSiteId() {
+  public Integer getSiteId() {
     return siteId;
   }
 
-  public void setSiteId(Long siteId) {
+  public void setSiteId(Integer siteId) {
     this.siteId = siteId;
   }
