+++ src/main/java/com/dt/user/model/SalesAmazonAd/SalesAmazonFbaBusinessreport.java
@@ -7,8 +7,8 @@ public class SalesAmazonFbaBusinessreport {
 
     private Long id;
     private Long date;
-    private Long shopId;
-    private Long siteId;
+    private Integer shopId;
+    private Integer siteId;
     private String sku;
     private Long skuId;
     private String fAsin;
@@ -38,7 +38,7 @@ public class SalesAmazonFbaBusinessreport {
 
     }
 
-    public SalesAmazonFbaBusinessreport(Long shopId, Long siteId, Long createDate, Long createIdUser, Long recordingId) {
+    public SalesAmazonFbaBusinessreport(Integer shopId, Integer siteId, Long createDate, Long createIdUser, Long recordingId) {
         this.shopId = shopId;
         this.siteId = siteId;
         this.createDate = createDate;
@@ -62,19 +62,19 @@ public class SalesAmazonFbaBusinessreport {
         this.date = date;
     }
 
-    public Long getShopId() {
+    public Integer getShopId() {
         return shopId;
     }
 
-    public void setShopId(Long shopId) {
+    public void setShopId(Integer shopId) {
         this.shopId = shopId;
     }
 
-    public Long getSiteId() {
+    public Integer getSiteId() {
         return siteId;
     }
 
-    public void setSiteId(Long siteId) {
+    public void setSiteId(Integer siteId) {
         this.siteId = siteId;
     }
