+++ src/main/java/com/dt/user/model/SalesAmazonAd/SalesAmazonFbaReceivestock.java
@@ -7,8 +7,8 @@ public class SalesAmazonFbaReceivestock {
 
   private Long id;
   private Long date;
-  private Long shopId;
-  private Long siteId;
+  private Integer shopId;
+  private Integer siteId;
   private String fbaShipmentId;
   private String sku;
   private String fnsku;
@@ -31,7 +31,7 @@ public class SalesAmazonFbaReceivestock {
 
   }
 
-  public SalesAmazonFbaReceivestock(Long shopId, Long createDate, Long createIdUser, Long recordingId) {
+  public SalesAmazonFbaReceivestock(Integer shopId, Long createDate, Long createIdUser, Long recordingId) {
     this.shopId = shopId;
     this.createDate = createDate;
     this.createIdUser = createIdUser;
@@ -55,19 +55,19 @@ public class SalesAmazonFbaReceivestock {
     this.date = date;
   }
 
-  public Long getShopId() {
+  public Integer getShopId() {
     return shopId;
   }
 
-  public void setShopId(Long shopId) {
+  public void setShopId(Integer shopId) {
     this.shopId = shopId;
   }
 
-  public Long getSiteId() {
+  public Integer getSiteId() {
     return siteId;
   }
 
-  public void setSiteId(Long siteId) {
+  public void setSiteId(Integer siteId) {
     this.siteId = siteId;
   }
